refactor: убирает директорию для констант
