+++ src/entities/city/const.ts
