+++ src/features/map/hooks/useMap.ts
@@ -1,6 +1,6 @@
 import { Map, TileLayer } from 'leaflet';
 import { useEffect, useState, MutableRefObject, useRef } from 'react';
-import { MAP_ZOOM, TileLayerSetup } from '../const/const';
+import { MAP_ZOOM, TileLayerSetup } from '../const';
 
 export function useMap(
   mapRef: MutableRefObject<HTMLElement | null>,
