+++ src/features/map/hooks/useMapPins.ts
@@ -1,7 +1,7 @@
 import { Icon, Marker, layerGroup, Map, BaseIconOptions } from 'leaflet';
 import { useEffect } from 'react';
 import 'leaflet/dist/leaflet.css';
-import { IconSetup } from '../const/const';
+import { IconSetup } from '../const';
 
 export function useMapPins(
   map: Nullable<Map>,
