+++ src/shared/lib/react/components/privateRoute.tsx
@@ -1,6 +1,6 @@
 import { JSX } from 'react';
 import { Navigate } from 'react-router-dom';
-import { AuthorizationStatus } from '../const/const';
+import { AuthorizationStatus } from '../const';
 
 type PrivateRouteProps = {
   authorizationStatus: AuthorizationStatus;
