+++ src/shared/lib/react/const.ts
