+++ src/shared/ui/starsRatingInfo/const.ts
