+++ src/shared/ui/starsRatingInfo/lib/getWidthByStarsCount.ts
@@ -1,3 +1,3 @@
-import { TOTAL_STARS_COUNT } from '../const/const';
+import { TOTAL_STARS_COUNT } from '../const';
 
 export const getWidthByStarsCount = (starsCount: number) => `${starsCount / TOTAL_STARS_COUNT * 100}%`;
