+++ src/widgets/nearPlaces/const.ts
