+++ src/widgets/offer/const.ts
