+++ src/widgets/offer/ui/offer.tsx
@@ -5,7 +5,7 @@ import { Map } from '@/features/map';
 import { Review } from '@/entities/review';
 import { capitalizeWord } from '@/shared/lib';
 import { Badge, StarsRatingInfo } from '@/shared/ui';
-import { MAX_IMAGES_COUNT } from '../const/const';
+import { MAX_IMAGES_COUNT } from '../const';
 import { reviews } from '@/mock/reviews';
 
 type OfferProps = {
